pyproject.toml
[tool.poetry]
name = "happybirthdaybot"
version = "0.1.0"
description = ""
authors = ["Norilka <NorilkaRus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
telebot = "^0.0.5"
aiogram = "^3.11.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
